requirements.txt
Pillow==4.2.1
Django==1.11.4